Split subite.js into individual functions
