+++ package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "test": "npm run lint && npm run ava",
     "ava": "ava test/*.test.js",
-    "lint": "standard subtitle.js test/*.test.js",
+    "lint": "standard subtitle.js lib/*.js test/*.test.js",
     "coverage": "nyc npm test",
     "report": "nyc report --reporter=html",
     "coveralls": "nyc report --reporter=text-lcov | coveralls",
